accounting_core::ledger::core

Struct Ledger

pub struct Ledger<S: LedgerStorage> { /* private fields */ }

Main ledger system that orchestrates all accounting operations

Implementations

impl<S: LedgerStorage + Clone> Ledger<S>

pub fn new(storage: S) -> Self

Create a new ledger with the given storage backend

pub fn with_validators( storage: S, account_validator: Box<dyn AccountValidator>, transaction_validator: Box<dyn TransactionValidator>, ) -> Self

Create a new ledger with custom validators

pub async fn create_account( &mut self, id: String, name: String, account_type: AccountType, parent_id: Option<String>, ) -> LedgerResult<Account>

Create a new account

pub async fn get_account( &self, account_id: &str, ) -> LedgerResult<Option<Account>>

Get an account by ID

pub async fn list_accounts(&self) -> LedgerResult<Vec<Account>>

List all accounts

pub async fn list_accounts_by_type( &self, account_type: AccountType, ) -> LedgerResult<Vec<Account>>

List accounts by type

pub async fn update_account(&mut self, account: &Account) -> LedgerResult<()>

Update an account

pub async fn delete_account(&mut self, account_id: &str) -> LedgerResult<()>

Delete an account

pub async fn record_transaction( &mut self, transaction: Transaction, ) -> LedgerResult<()>

Record a new transaction

pub async fn get_transaction( &self, transaction_id: &str, ) -> LedgerResult<Option<Transaction>>

Get a transaction by ID

pub async fn get_account_transactions( &self, account_id: &str, start_date: Option<NaiveDate>, end_date: Option<NaiveDate>, ) -> LedgerResult<Vec<Transaction>>

Get transactions for a specific account

pub async fn get_transactions( &self, start_date: Option<NaiveDate>, end_date: Option<NaiveDate>, ) -> LedgerResult<Vec<Transaction>>

Get all transactions within a date range

pub async fn update_transaction( &mut self, transaction: &Transaction, ) -> LedgerResult<()>

Update a transaction

pub async fn delete_transaction( &mut self, transaction_id: &str, ) -> LedgerResult<()>

Delete a transaction

pub async fn get_account_balance( &self, account_id: &str, as_of_date: Option<NaiveDate>, ) -> LedgerResult<BigDecimal>

Get account balance as of a specific date

pub async fn get_trial_balance( &self, as_of_date: NaiveDate, ) -> LedgerResult<TrialBalance>

Get trial balance as of a specific date

pub async fn get_account_balances_by_type( &self, as_of_date: NaiveDate, ) -> LedgerResult<HashMap<AccountType, Vec<AccountBalance>>>

Get account balances grouped by type

pub async fn generate_balance_sheet( &self, as_of_date: NaiveDate, ) -> LedgerResult<BalanceSheet>

Generate a balance sheet as of a specific date

pub async fn generate_income_statement( &self, start_date: NaiveDate, end_date: NaiveDate, ) -> LedgerResult<IncomeStatement>

Generate an income statement for a date range

pub async fn generate_cash_flow( &self, start_date: NaiveDate, end_date: NaiveDate, ) -> LedgerResult<CashFlowStatement>

Create a basic cash flow statement

pub async fn setup_standard_chart_of_accounts( &mut self, ) -> LedgerResult<HashMap<String, Account>>

Setup a standard chart of accounts for small business

pub async fn validate_integrity( &self, as_of_date: NaiveDate, ) -> LedgerResult<LedgerIntegrityReport>

Validate the integrity of the ledger